move universal autolog tests to different file and mark as large
